🚧 2NHAWN task: split harvest artifacts module
Task: 202605131049-2NHAWN
